requirements.txt
praw
attrs
pyyaml
python-benedict
streamlit
openai